pyproject.toml
[tool.poetry]
name = "dotenv-manager"
version = "0.1.0"
description = ""
authors = ["Teodor Toshkov <teodor.toshkov95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
